+++ .hooks/pre-commit
@@ -5,14 +5,8 @@ echo "Validating files"
 python Tests/scripts/validate_files_structure.py
 
 if [[ $? -ne 0 ]]
-  then
     echo "Please fix the aforementioned errors and then commit again"
     exit 1
-
-  else
-    echo "Updating id_set"
-    python Tests/scripts/update_id_set.py
-    git add Tests/id_set.json
 fi
